+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - `appengine devices send-data` now parses integer values correctly for server-owned datastream aggregate
   interfaces
+- `appengine devices data-snapshot` now handles partial failures in a better way without compromising the
+  full results
 
 ## [0.11.0-rc.1] - 2020-04-01
 ### Added